[REF] web, mail, *: move html utils to web
* = website_forum

Having 2 diffrent html utils files is confusing and creates dependency
issues for modules that need to use the HTML utils currently in mail if
they don't have mail as dependency (web itself and the various html
editor modules in particular).

More and more PR in these modules could make use of the mail utils
(htmlReplace in particular, but not limited to it) and they currently
repeat its code instead of using it due to the dependency issue.

This commit purposely doesn't make use of the utils where they could be
used to reduce its scope and avoid conflicts with other PRs.
