+++ addons/mail/__manifest__.py
@@ -174,7 +174,6 @@ For more specific needs, you may also assign custom-defined actions
             'mail/static/src/**/*.dark.scss',
         ],
         "web.assets_frontend": [
-            "mail/static/src/utils/common/html.js",
             "mail/static/src/utils/common/format.js",
         ],
         'mail.assets_discuss_public_test_tours': [
