+++ addons/mail/static/src/chatter/web/form_controller.js
@@ -1,9 +1,8 @@
-import { createDocumentFragmentFromContent } from "@mail/utils/common/html";
-
 import { EventBus, useSubEnv } from "@odoo/owl";
 
 import { x2ManyCommands } from "@web/core/orm_service";
 import { useService } from "@web/core/utils/hooks";
+import { createDocumentFragmentFromContent } from "@web/core/utils/html";
 import { patch } from "@web/core/utils/patch";
 import { FormController } from "@web/views/form/form_controller";
