+++ addons/mail/static/src/core/web/composer_patch.js
@@ -3,10 +3,10 @@ import { wrapInlinesInBlocks } from "@html_editor/utils/dom";
 import { childNodes } from "@html_editor/utils/dom_traversal";
 
 import { Composer } from "@mail/core/common/composer";
-import { createDocumentFragmentFromContent } from "@mail/utils/common/html";
 
 import { markup } from "@odoo/owl";
 
+import { createDocumentFragmentFromContent } from "@web/core/utils/html";
 import { patch } from "@web/core/utils/patch";
 import { renderToElement } from "@web/core/utils/render";
